Update the API re `PostMessageOptions` vs `StructuredSerializeOptions`
whatwg/html#3414 introduced the `structuredClone()` API and renamed the
`PostMessageOptions` dictionary to `StructuredSerializeOptions`. When
implementing this change in Chromium
(https://chromium-review.googlesource.com/c/chromium/src/+/3245334), it
was decided that, since `structuredClone()` has no use for the
`includeUserActivation` parameter, it was better to keep the
`StructuredSerializeOptions` as per the spec and incorporate that
parameter in a derived interface keeping the old name
`PostMessageOptions`.

This change updates the proposed IDL of the explainer to reflect these
changes.